MN-15 refactor
